0506 16:11 kjr 디테일 별점평균, 장바구니에 넣기 수정

+++ sts0417/src/main/java/net/bs/spring/BookDAO.java
@@ -100,7 +100,7 @@ public class BookDAO {
 		return temp.selectOne("books.select_detail", idx);
 	}
 	
-	public int dbStaravg(int idx) {   //디테일 별점 평균
+	public double dbStaravg(int idx) {   //디테일 별점 평균
 		return temp.selectOne("books.star_avg", idx);
 	}
